fixed #2 and fixed #3
The Window is now centered in the middle of the screen.

Added the README.

Added Field class.

Reworked most of the Model class.

Added methods to set size and bombCount (Still no UI to set them).

Added moveBomb method to move bombs, thereby allowing the first move to
be bombfree and as preparation for future features.

Fixed a bug that caused fill to reveal the top left field if there was a
zero nearby.

Removed unnecessary/invalid methods and comments.